desktop-tools: remove vpn setup
This is also my last commit using semantic commits. I find it not very useful for this repository.
It might be more useful to others where I would like to generate automatic releases or metadata out
of vcs history.

For this use-case, I find it much easier to write AND to understand what a commit has ever affected.
This is the same commit structure used by the kernel, so it's very battle tested and in my own
experience, best suitable for my dotfiles :)