zsh: add support for loading dotenv on non-interactive sessions
This allows us to use `zsh -c` to launch any programs that were previously only accessible through
an ZSH interactive session, such as Emacs, which is only able to load up few dependencies when it
knows where to look for (because I touch with $PATH on my dotfiles, it wouldn't ever know until it
loaded them).

Also, adds an overriden `emacs.desktop` app launcher that uses the zsh as the wrapper to load the
dotenv environment beforehand.